+++ pkg/wharfyml/magicstrings.go
@@ -1,6 +1,10 @@
 package wharfyml
 
 const (
+	// Map keys in .wharf-ci.yml
 	propInputs       = "inputs"
 	propEnvironments = "environments"
+
+	// Map keys in .wharf-vars.yml
+	propVars = "vars"
 )